+++ CHANGES.md
@@ -21,7 +21,7 @@ Changes:
 -
 
 Fixes:
--
+- Fixed the `nh_p` and `nh_u` scrapers #1663
 
 ## 3.0.1 - 2026-03-18