+++ CHANGES.md
@@ -15,11 +15,21 @@ Releases are also tagged in git, if that's helpful.
 The following changes are not yet released, but are code complete:
 
 Features:
+-
+
+Changes:
+-
+
+Fixes:
+-
+
+## 3.0.2 - 2026-03-20
+
+Features:
 - Implements `kan` and `kanctapp` backscraper #1316
 
 Changes:
 - `kan` back to OpinionSiteLinear
-
 
 Fixes:
 - Fixes `kan` and `kanctapp` by using urllib requests #1316